% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converttime.R
\name{converttime}
\alias{converttime}
\title{Convert time from seconds into readable format}
\usage{
converttime(number)
}
\arguments{
\item{number}{is a value in seconds}
}
\value{
The vale in Day:Hours:Minutes:Seconds format
}
\description{
Convert time from seconds into readable format
}
