% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradata.R
\name{count_answer_changes}
\alias{count_answer_changes}
\title{Count the number of answer changes per interview}
\usage{
count_answer_changes(df)
}
\arguments{
\item{df}{Data frame that is Survey Solutions' paradata}
}
\value{
Data frame.
}
\description{
Count the number of answer changes per interview
}
