% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradata.R
\name{find_paradata}
\alias{find_paradata}
\title{Find paradata files}
\usage{
find_paradata(dir, file_pattern = "paradata\\\\.tab", recurse = TRUE)
}
\arguments{
\item{dir}{Character. Directory to scan for paradata files}

\item{file_pattern}{Character. Default value is SuSo's default file name.}

\item{recurse}{Boolean. Default value assumes that \code{dir} is parent diectory
that contains child directories that contain the paradata.}
}
\value{
Character vector of paths to paradata file(s).
}
\description{
Return path(s) of paradata files found target path.
}
