% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviewTable.R
\name{interviewTable}
\alias{interviewTable}
\title{interviewTable}
\usage{
interviewTable(x)
}
\arguments{
\item{x}{This must be the interview actions file in .tab format.}
}
\value{
A dataframe with the computations listed above.
}
\description{
This function requires the interview actions as a dataframe. 
It returns a dataframe sorted by interview including interview id,
Interviewer, Supervisor, HQApproved (# of times), SuperApproved (# of times)
Starttime, Endtime, and Duration (FirstAnswerSet - Completed)
}
\author{
Michael Rahija
}
