% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradata.R
\name{parse_paradata}
\alias{parse_paradata}
\title{Parse paradata file}
\usage{
parse_paradata(dt)
}
\arguments{
\item{dt}{Data table or data frame}
}
\value{
Data table of pradata with three new columns: \code{variable}, \code{value}, and \code{row}.
}
\description{
Transforms paradata into a more usable format by splitting the \code{parameters} into its constituent pieces: \code{variable}, \code{value}, and \code{row}
}
