% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradata.R
\name{read_paradata}
\alias{read_paradata}
\title{Read pardata file(s) from disk}
\usage{
read_paradata(file)
}
\arguments{
\item{file}{Character vector. One or more full path to paradata file(s).}
}
\value{
Data table
}
\description{
Reads paradata file(s) using a fast method
}
