% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradata.R
\name{summarize_answer_changes}
\alias{summarize_answer_changes}
\title{Summarize answer changes}
\usage{
summarize_answer_changes(df)
}
\arguments{
\item{df}{Data frame that is Survey Solutions' paradata}
}
\value{
Data frame.
}
\description{
Produces a data frame that summarizes both the frequency and details of answer changes.
An answer change may involve one, many, or a combination of the following interviewer actions: answer changed, answer removed.
The summary reduces the answer change events for a question into a formatted string of the following format:
\verb{(order) (+/-) value}, where \code{order} is the event order number in the paradata,
\verb{(+/-)} indicates an addition/change (\verb{(+)}) or a deletion (\verb{(-)}),
and \code{value} shows answer value(s) due to that event.
The summary will consist of several such event blocks separated by a semi-colon (i.e., \verb{;}).
}
