% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradata.R
\name{calc_time_btw_active_events}
\alias{calc_time_btw_active_events}
\title{Calculate time between active paradata events}
\usage{
calc_time_btw_active_events(df)
}
\arguments{
\item{df}{Data frame that is Survey Solutions' paradata}
}
\value{
Data frame. Adds a few columns to the input data set: event duration in elapsed seconds (\code{elapsed_sec}) and minutes (\code{elapsed_min}). Removes passive events.
}
\description{
Determines active event durations by:
}
\details{
\itemize{
\item Filtering out passive and non-interview events
\item Computing difference between active events
}

Passive events include:
\itemize{
\item Events of the following types in \code{event}: c("ApproveByHeadquarter", "ApproveBySupervisor", "ClosedBySupervisor", "KeyAssigned", "OpenedBySupervisor", "QuestionDeclaredInvalid", "QuestionDeclaredValid", "ReceivedByInterviewer", "ReceivedBySupervisor", "RejectedByHeadquarter", "RejectedBySupervisor", "SupervisorAssigned", "TranslationSwitched", "UnapproveByHeadquarters", "VariableDisabled", "VariableSet")
\item Pauses in the interview--that is, time between \code{Pause} and \code{Resume} events
\item End and resumption of the interview--that is between events and \code{Restarted} or between \code{Restarted} and an event
}

All other events are considered active
}
