% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradata.R
\name{read_paradata}
\alias{read_paradata}
\title{Read pardata from disk}
\usage{
read_paradata(file)
}
\arguments{
\item{file}{Full file path to the paradata file.}
}
\value{
Data table
}
\description{
Reads paradata using a fast method
}
